+++ crates/task/src/task.rs
@@ -12,13 +12,19 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::env;
 use std::path::PathBuf;
+use strum::Display;
 
-#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, Display, Eq, PartialEq, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub enum TaskType {
+    #[strum(serialize = "build")]
     Build,
+
+    #[strum(serialize = "run")]
     Run,
+
     #[default]
+    #[strum(serialize = "test")]
     Test,
 }
 
@@ -240,7 +246,7 @@ impl Task {
             },
             outputs: cloned_config.outputs.unwrap_or_default(),
             output_paths: HashSet::new(),
-            platform: cloned_config.type_of,
+            platform: cloned_config.platform,
             target: target.id.clone(),
             type_of: TaskType::Test,
         };
@@ -284,7 +290,7 @@ impl Task {
         }
 
         if !matches!(self.platform, PlatformType::Unknown) {
-            config.type_of = self.platform.clone();
+            config.platform = self.platform.clone();
         }
 
         config
@@ -581,7 +587,7 @@ impl Task {
 
         // Merge options first incase the merge strategy has changed
         self.options.merge(&config.options);
-        self.platform = config.type_of.clone();
+        self.platform = config.platform.clone();
 
         // Then merge the actual task fields
         if let Some(cmd) = command {